refactor: model 내 클래스를 domain과 service 패키지로 분리
