+++ src/main/java/lotto/model/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.ArrayList;
 import java.util.List;
